Add copyright headers to various files.
I'd like to package sunxi-tools for Debian and therefore it is important for
the licensing information to be complete/accurate. I believe the intention was
for everything here to be GPL2+ by default, since that is the license on every
file which has one and COPYING contains GPL2.

Early on the license applied to this repo was GPLv3 however this was changed to
GPL2+ by Alejandro in 79ea14d4e050 at which point he had been the only
contributor.

This patch adds the standard GPL2+ stanza used already in sunxi-tools.git or
the MIT license stanza when requested by the copyright holder to various files
which were missing one as follows:

adb-devprobe.sh
fel-gpio
    According to git all of these were written by Henrik. Copyright years
    according to git. Henrik requested that these be put under an MIT license,
    so that is what has been done.

boot_head.lds
fel-pio.lds
fel-sdboot.lds
jtag-loop.lds
    According to git all of these were written by Henrik. Copyright years
    according to git. According to Henrik "These linker scripts are all GPLv2+
    as the C / ASM sources they refer to".

include/endian_compat.h
    The content of this file was originally added to fel.c (commit
    c71ff92c02d3), which had a GPL2+ stanza at the time, by Eric Molitor and later
    those lines were moved by Alejandro (commit bcde0fc72170) into this file.
    I originally added GPL2+ from fel.c and added Eric's copyright with the
    correct year according to git but Eric said "Ack but also would prefer
    MIT/Dual :)", so it has now been changed to MIT.

include/types.h:
    Henrik originally added some of these lines to bootinfo.c, along with a
    GPL2+ stanza, in the original version (commit c26e5ff80af6). Later on
    Alejandro moved them into this file (commit 329a13ed75e5) and added more.
    I've copied the stanza from bootinfo.c and added both copyrights with the
    years according to git. Authors:
      Henrik, who says "OK".
      Alejandro Mery

Makefile:
    GPL2+ with copyrights and years according to git. Authors are:
      Alejandro Mery
      Henrik Nordstrom, who says "Yes"
      Pat Wood, who says "Fine with me"

usb-boot:
    Henrik is the primary author, added MIT license on Henrik's request with
    his copyright and years according to git. Authors are:
      Henrik Nordstrom, who says "Yes, that too should be MIT"
      Alejandro Mery (typo fix)
      Michal Suchanek (typo fix)

Everyone affected by the above is CCd.

This probably seems pretty obvious to most people, sorry for being so pedantic
about it. It will save hassel when it comes to getting it into Debian though.

Signed-off-by: Ian Campbell <ijc@hellion.org.uk>
Cc: Henrik Nordstrom <henrik@henriknordstrom.net>
Cc: Eric Molitor <eric@molitor.org>
Cc: Alejandro Mery <amery@geeks.cl>
Cc: Pat Wood <Pat.Wood@efi.com>
Cc: Michal Suchanek <hramrach@gmail.com>
---
v2: Gathered feedback from the authors
 - Pat said "Fine with me"
 - Henrik asked that adb-devprobe.sh, fel-gpio and usb-boot be MIT, acked
   *.lds, types.h and Makefile
 - Explicitly listed authors of Makefile and usb-boot
 - Michal Suchanek make a typo fix to usb-boot but wasn't CCd, sorry.
 - Reworded commit message due to some bits now being MIT on request of the
   author.