fel-to-spl-thunk.S: Only depend on binutils.
A cross-binutils is generally a bit easier to get hold of, and in particular
binutils-none-eabi (and binutils-arm-linux-gnueabi{,hf}) are available in
Debian Jessie while the equivalent cross gcc is not.

The only reason gcc/cpp are required are to strip the embedded Ruby code out
before handing to the assembler, we can achieve the same by opening a multiline
comment around the ruby instead. Care needs to be taken not to close the
comment prematurely hence "*/" is written in the one place it is used as
"\x2a/" (i.e. encoding the * in hex). Having done this we can pass the .S file
directly to the cross-as.

There is no change to the resulting header file.

Signed-off-by: Ian Campbell <ijc@hellion.org.uk>
Acked-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>