fel: Handle error and don't exit if MMU is not enabled.
On allwinner SOCs A33,H3, A83T MMU is not enabled by BROM, so don't exit
if MMU is not enabled by BROM and return NULL.
It was reported on A33 tablet and Orange Pi H3 Board.
I've tested it on A33 tablet and A83T Dev Board by allwinner.

Signed-off-by: VishnuPatekar <vishnupatekar0510@gmail.com>
Acked-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>