fel-to-spl-thunk.S: Fix safety guard which checks for enabled cache
The exit code path after detecting that the cache is enabled does not
need to swap FEL stacks and backup buffers. This way we get the error
actually reported by the 'fel' tool instead of the device getting
locked up.

The thunk code refuses to work if the caches are enabled because
we don't want to deal with the instructions/data cache coherency
yet. The caches are initially in a disabled state upon activating
FEL mode on all currently known Allwinner processors.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>
Acked-by: Hans de Goede <hdegoede@redhat.com>