README: Remove the obsolete 'usb-boot' script
It was designed to be used with the legacy u-boot-sunxi
bootloader and has many assumptions, which are specific to
it (a special FEL configuration, hardcoded addresses, etc.)

The mainline U-Boot can now be booted over USB with the
"fel uboot u-boot-sunxi-with-spl.bin" command. And this
makes the old 'usb-boot' script obsolete.

The presence of the 'usb-boot' script in the repository may
confuse the users, so we are removing it now. The users of the
old u-boot-sunxi bootloader still can always get this script
from one of the old sunxi-tools tags.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>
Acked-by: Hans de Goede <hdegoede@redhat.com>