Makefile: Prefix most tools with sunxi-
Several of the tools here are too generic and/or short for distro packaging
purposes (which like to try and avoid naming clashes in $PATH). Prefix the
following with "sunxi-":
  - fexc
  - bootinfo
  - fel
  - pio
  - meminfo

Do not prefix any of the fel "payloads" or raw binaries since they would not
normally be installed in $PATH.

Do not prefix "phoenix_info", since "phoenix" already seems like an appropriate
prefix for this particular tool.

Update in-tree callers, README and .gitignore accordingly.

Signed-off-by: Ian Campbell <ijc@hellion.org.uk>
Acked-by: Peter Korsgaard <peter@korsgaard.com>