Makefile: Split out target tools rules
Target tools are those which are only useful on a target sunxi system (i.e.
which probe hardware etc).

Currently this is only sunxi-pio. At first I thought sunxi-nand-part might be
included, but I think that is useful on NAND images as well as actual devices.

This will allow for easier packaging, by letting packagers only include the
target tools when building for a suitable ARM architecture.

Signed-off-by: Ian Campbell <ijc@hellion.org.uk>
Acked-by: Peter Korsgaard <peter@korsgaard.com>