sunxi-tools: implement simple progress bar for larger transfers
Add "--progress" option and a progress bar display for FEL
transfers. This picks up on a suggestion from Alexander Kaplan
and the discussion at
https://groups.google.com/forum/#!topic/linux-sunxi/lz0oQBwjex0

Note that this particular progress callback - and thus the
display output - will ignore any smaller transfers, i.e. those
that complete in a single loop iteration. In other words:
progress will only be shown if the transfer size requires
multiple chunks to be sent.

Signed-off-by: Bernhard Nortmann <bernhard.nortmann@web.de>