fel: support updating the caption for "dialog --gauge"
This patch adds an "echo-gauge" command that is intended for use with
sunxi-fel's output piped into the dialog utility. "echo-gauge" will
output its string argument in a way that updates (i.e. overwrites)
the prompt text that dialog displays, so it's possible to change that
inbetween file transfers from sunxi-fel.

Example:
	sunxi-fel uboot ${UBOOT} \
		echo-gauge "Uploading kernel (1/2)" \
		write-with-gauge 0x42000000 ${KERNEL} \
		write 0x43100000 ${SCRIPT} \
		echo-gauge "Uploading rootfs (2/2)" \
		write-with-gauge 0x44000000 ${ROOTFS} \
		| dialog --gauge "" 6 70

Signed-off-by: Bernhard Nortmann <bernhard.nortmann@web.de>
Reviewed-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>