script_extractor: Remove unnecessary size argument
The script_extractor tool before this commit used to take a size argument
on the cmdline, but the passed in size was only used in some places not
in others. Leading to a segfault if the passed in argument was not
exactly the same as SCRIPT_SIZE.

This commit drops the argument, so that script_extractor will just
work.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>