fel-sdboot.sunxi: Add support for A64 and A80
This small bootable stub, which just passes control to the
FEL code in the BROM, needs to be adjusted to also support
Allwinner A64 and Allwinner A80 because the BROM is located
at a different address there.

The SD card boot has a very low priority on Allwinner A64, but
it at least has a higher priority than the SPI NOR Flash:

    https://linux-sunxi.org/BROM#A64

So this patch may help to simplify the FEL mode activation on
devices, which are booting their firmware from the SPI NOR Flash.

Changes in v2:
 - Use SCTLR.V to detect the exception vectors location as
   suggested by Jens Kuske.
 - Add a padding of 32 NOP instructions in the beginning as
   a workaround for some strange failures.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>