fexc: Warn when decompiling a malformed section entry (key string)
script_bin.c decompiles section entries even if they have invalid
indentifiers. Thus malformed .bin files were observed to result in
a .fex that couldn't be processed successfully with the sunxi-fexc
compiler.

This patch makes bin2fex now issue a warning for this kind of keys,
so that they're easier to notice and correct.

Signed-off-by: Bernhard Nortmann <bernhard.nortmann@web.de>