travisci: Adjust build config to use "make all" on Linux
For Linux build testing, we want all targets compiled. To do so
without a cross-toolchain, simply use the host compiler for the
target tools. This can be achieved by setting CROSS_COMPILE to
an empty string.

OSX can't handle this: It neither supports "-static", nor would
it successfully compile meminfo.c. Thus we keep the default
"make", which should only build the 'tools' target.

On Linux we'll also test the install-* targets by requesting
installation to /tmp/sunxi-tools as a post-build step.

Signed-off-by: Bernhard Nortmann <bernhard.nortmann@web.de>