fel: Parse SPL DT name
A while ago the SPL header was extended to hold the name of the DTB file
that shall be used for the board a firmware image is made for.

Add some code to extract that name from the SPL header. This will be
used in later patches to load the proper DTB file.

Signed-off-by: Andre Przywara <osp@andrep.de>