fel: Check for U-Boot image before considering checksum
Currently we check the U-Boot (legacy!) image header checksum very early
and bail out with an error message if it does not match.

Move that check later into the function, *after* we have established
that we are actually dealing with such an U-Boot image.

This avoids confusing error messages in case there is no U-Boot image
used at all.

Signed-off-by: Andre Przywara <osp@andrep.de>