fel: thunks: Fix fel-to-spl-thunk to be ARMv5TE compatible
Currently the thunk we upload into the SRAM is using DSB and ISB
instructions, which were introduced in ARMv7. Also it relies on
movw/movt pairs, which became available in ARMv6T2.
The Allwinner F1Cx00 SoCs are using an ARMv5TE compliant core, so they
do not know these instructions.

Change the code to be ARMv5TE compliant, so it can run on all relevant
Allwinner ARM cores:
- One movw is just used to compare two bits, replace that with a
  tst/tsteq sequence to skip the load.
- The other movw/movt pairs get replaced with ldr's, that load from
  literal storage at the end of the code (from Icenowy).
- The DSB and ISB get replaced with their CP15 MCR counterparts. Those
  are deprecated in ARMv7, but still work, when the CP15BEN bit is set
  in SCTLR. We check for this in fel.c (from Icenowy). ISB is not
  implemented on the ARM926, so make this conditional. A simple branch
  takes care of the desired pipeline flush for the old SoC.

Also remove the rather pointless Ruby prolog that generates the header
file. We have a less awkward version of this in the Makefile, and need
that for the other thunks there anyway, so it's just duplicated code.
Embedding a header generator in Ruby in an assembly file is a cute
gimmick, but serves no purpose anymore.

This is based on work by Icenowy, who put a similar solution in a
separate file.

Originally-by: Icenowy Zheng <icenowy@aosc.io>
Signed-off-by: Andre Przywara <osp@andrep.de>