CI: add Github Actions control file
It seems like travis-ci.org is gone, replaced by travis-ci.com, but
apparently that's not truly free anymore: it requires a registered account
plus credit card, and is only granted a one-time free credit allowance,
with the option to get this extended by explicit request. As a result
the current CI stopped working a while ago, and was broken anyway (it
failed building already, and make check on its own doesn't pass on
master anymore).

Github itself seems to offer an alternative, so let's add a control file
for that to get some basic CI back.
This builds the host and misc tools on a standard Ubuntu system, and
runs make check.

Signed-off-by: Andre Przywara <osp@andrep.de>