tests: fex: remove remaining patch file
The sunxi-boards.git repository has been fixed to address the two
remaining issues in the .fex files there, so we can run our fex compiler
tests on the unpatched repository content.

Remove the patch file, as it is no longer needed and won't apply
anymore anyway on the current sunxi-boards master.

Signed-off-by: Andre Przywara <osp@andrep.de>